refactor codebase to lean on Role model methods
