+++ repokid/dispatcher/__init__.py
@@ -7,10 +7,12 @@ from mypy_boto3_dynamodb.service_resource import Table
 
 import repokid.commands.repo
 import repokid.utils.dynamo as dynamo
+import repokid.utils.permissions
 import repokid.utils.roledata as roledata
 from repokid import CONFIG
 from repokid import get_hooks
 from repokid.dispatcher.types import Message
+from repokid.role import Role
 
 ResponderReturn = namedtuple("ResponderReturn", "successful, return_message")
 
@@ -48,18 +50,17 @@ def list_repoable_services(dynamo_table: Table, message: Message) -> ResponderRe
             ),
         )
     else:
-        role_data = dynamo.get_role_data(
-            dynamo_table, role_id, fields=["RepoableServices"]
-        )
+        role = Role(role_id=role_id)
+        role.fetch(fields=["RepoableServices"])
 
         (
             repoable_permissions,
             repoable_services,
         ) = roledata._convert_repoed_service_to_sorted_perms_and_services(
-            set(role_data.repoable_services)
+            set(role.repoable_services)
         )
 
-        repoable_services = role_data.repoable_services
+        repoable_services = role.repoable_services
         return ResponderReturn(
             successful=True,
             return_message=(
@@ -86,22 +87,23 @@ def list_role_rollbacks(dynamo_table: Table, message: Message) -> ResponderRetur
                 message.role_name, message.account
             ),
         )
-    else:
-        role_data = dynamo.get_role_data(dynamo_table, role_id, fields=["Policies"])
-        return_val = "Restorable versions for role {} in account {}\n".format(
-            message.role_name, message.account
+
+    role = Role(role_id=role_id)
+    role.fetch(fields=["Policies"])
+    return_val = "Restorable versions for role {} in account {}\n".format(
+        message.role_name, message.account
+    )
+    for index, policy_version in enumerate(role.policies):
+        total_permissions, _ = repokid.utils.permissions.get_permissions_in_policy(
+            policy_version["Policy"]
         )
-        for index, policy_version in enumerate(role_data.policies):
-            total_permissions, _ = roledata.get_permissions_in_policy(
-                policy_version["Policy"]
-            )
-            return_val += "({:>3}):  {:<5}     {:<15}  {}\n".format(
-                index,
-                len(total_permissions),
-                policy_version["Discovered"],
-                policy_version["Source"],
-            )
-        return ResponderReturn(successful=True, return_message=return_val)
+        return_val += "({:>3}):  {:<5}     {:<15}  {}\n".format(
+            index,
+            len(total_permissions),
+            policy_version["Discovered"],
+            policy_version["Source"],
+        )
+    return ResponderReturn(successful=True, return_message=return_val)
 
 
 @implements_command("opt_out")
@@ -128,9 +130,10 @@ def opt_out(dynamo_table: Table, message: Message) -> ResponderReturn:
             ),
         )
 
-    role_data = dynamo.get_role_data(dynamo_table, role_id, fields=["OptOut"])
-    if role_data.opt_out:
-        timestr = time.strftime("%m/%d/%y", time.localtime(role_data.opt_out["expire"]))
+    role = Role(role_id=role_id)
+    role.fetch(fields=["OptOut"])
+    if role.opt_out:
+        timestr = time.strftime("%m/%d/%y", time.localtime(role.opt_out["expire"]))
         return ResponderReturn(
             successful=False,
             return_message=(
@@ -138,8 +141,8 @@ def opt_out(dynamo_table: Table, message: Message) -> ResponderReturn:
                 "until {}".format(
                     message.role_name,
                     message.account,
-                    role_data.opt_out["owner"],
-                    role_data.opt_out["reason"],
+                    role.opt_out["owner"],
+                    role.opt_out["reason"],
                     timestr,
                 )
             ),
@@ -153,7 +156,8 @@ def opt_out(dynamo_table: Table, message: Message) -> ResponderReturn:
             "reason": message.reason,
             "expire": expire_epoch,
         }
-        dynamo.set_role_data(dynamo_table, role_id, {"OptOut": new_opt_out})
+        role.opt_out = new_opt_out
+        role.store(fields=["opt_out"])
         return ResponderReturn(
             successful=True,
             return_message="Role {} in account {} opted-out until {}".format(
@@ -176,9 +180,10 @@ def remove_opt_out(dynamo_table: Table, message: Message) -> ResponderReturn:
             ),
         )
 
-    role_data = dynamo.get_role_data(dynamo_table, role_id, fields=["OptOut"])
+    role = Role(role_id=role_id)
+    role.fetch(fields=["OptOut"])
 
-    if not role_data.opt_out:
+    if not role.opt_out:
         return ResponderReturn(
             successful=False,
             return_message="Role {} in account {} wasn't opted out".format(
@@ -186,7 +191,8 @@ def remove_opt_out(dynamo_table: Table, message: Message) -> ResponderReturn:
             ),
         )
     else:
-        dynamo.set_role_data(dynamo_table, role_id, {"OptOut": {}})
+        role.opt_out = {}
+        role.store(fields=["opt_out"])
         return ResponderReturn(
             successful=True,
             return_message="Cancelled opt-out for role {} in account {}".format(
