+++ repokid/lib/__init__.py
@@ -34,7 +34,7 @@ from repokid.commands.role_cache import _update_role_cache
 from repokid.commands.schedule import _cancel_scheduled_repo
 from repokid.commands.schedule import _schedule_repo
 from repokid.commands.schedule import _show_scheduled_roles
-from repokid.utils.dynamo import dynamo_get_or_create_table
+from repokid.utils.dynamo_v2 import dynamo_get_or_create_table
 
 hooks = get_hooks(CONFIG.get("hooks", ["repokid.hooks.loggers"]))
 dynamo_table: Table = dynamo_get_or_create_table(**CONFIG["dynamo_db"])
