+++ repokid/role.py
@@ -14,12 +14,15 @@
 from __future__ import annotations
 
 import datetime
+import logging
 import time
 from typing import Any
 from typing import Dict
+from typing import Iterable
 from typing import List
 from typing import Optional
 from typing import Set
+from typing import Tuple
 from typing import Union
 from typing import overload
 
@@ -32,11 +35,18 @@ from repokid.exceptions import IntegrityError
 from repokid.exceptions import RoleModelError
 from repokid.exceptions import RoleNotFoundError
 from repokid.types import RepokidConfig
+from repokid.types import RepokidHooks
 from repokid.utils.aardvark import get_aardvark_data
 from repokid.utils.dynamo_v2 import create_dynamodb_entry
 from repokid.utils.dynamo_v2 import get_role_by_id
 from repokid.utils.dynamo_v2 import get_role_by_name
 from repokid.utils.dynamo_v2 import set_role_data
+from repokid.utils.permissions import _convert_repoable_perms_to_perms_and_services
+from repokid.utils.permissions import _get_repoable_permissions
+from repokid.utils.permissions import find_newly_added_permissions
+from repokid.utils.permissions import get_permissions_in_policy
+
+logger = logging.getLogger("repokid")
 
 
 def to_camel(string: str) -> str:
@@ -51,8 +61,8 @@ class Role(BaseModel):
     assume_role_policy_document: Dict[str, Any] = Field(default={})
     create_date: Optional[datetime.datetime] = Field()
     disqualified_by: List[str] = Field(default=[])
-    last_updated: datetime.datetime = Field(default_factory=datetime.datetime.now)
-    no_repo_permissions: Dict[str, Any] = Field(default={})
+    last_updated: Optional[datetime.datetime] = Field()
+    no_repo_permissions: Dict[str, int] = Field(default={})
     opt_out: Dict[str, int] = Field(default={})
     policies: List[Dict[str, Any]] = Field(default=[])
     refreshed: Optional[str] = Field()
@@ -76,6 +86,8 @@ class Role(BaseModel):
         "_dirty",
         "_updated_fields",
     }
+    # TODO: read exclude_new_permissions_for_days from config
+    _no_repo_secs: int = 24 * 60 * 60 * 14
     _updated_fields: Set[str] = PrivateAttr(default_factory=set)
 
     class Config:
@@ -92,6 +104,91 @@ class Role(BaseModel):
 
     def __repr__(self) -> str:
         return f"<Role {self.role_id}>"
+
+    def add_policy_version(
+        self, policy: Dict[str, Any], source: str = "Scan", store: bool = True
+    ) -> None:
+        if self.policies:
+            last_policy = self.policies[-1]["Policy"]
+            if policy == last_policy:
+                # we're already up to date, so this is a noop
+                return
+        else:
+            last_policy = {}
+        policy_entry = {
+            "Source": source,
+            "Discovered": datetime.datetime.utcnow().isoformat(),
+            "Policy": policy,
+        }
+        self.policies.append(policy_entry)
+        if store:
+            self.store(fields=["Policies", "NoRepoPermissions"])
+
+    def _calculate_repo_scores(self, minimum_age: int, hooks: RepokidHooks) -> None:
+        all_permissions, eligible_permissions = get_permissions_in_policy(
+            self.policies[-1].get("Policy", {})
+        )
+        self.total_permissions = len(all_permissions)
+        if self.disqualified_by or not self.aa_data:
+            self.repoable_permissions = 0
+            self.repoable_services = []
+            return
+        repoable_permissions = _get_repoable_permissions(
+            self.account,
+            self.role_name,
+            eligible_permissions,
+            self.aa_data,
+            self.no_repo_permissions,
+            self.role_id,
+            minimum_age,
+            hooks,
+        )
+        (
+            repoable_permissions_set,
+            repoable_services_set,
+        ) = _convert_repoable_perms_to_perms_and_services(
+            eligible_permissions, repoable_permissions
+        )
+        self.repoable_services = list(
+            repoable_permissions_set.union(repoable_permissions_set)
+        )
+        self.repoable_permissions = len(repoable_permissions)
+
+    def _get_permissions(
+        self, warn_unknown_perms: bool = False
+    ) -> Tuple[Set[str], Set[str]]:
+        if not self.policies:
+            return set(), set()
+
+        return get_permissions_in_policy(
+            self.policies[-1]["Policy"], warn_unknown_perms=warn_unknown_perms
+        )
+
+    def _update_no_repo_permissions(self) -> None:
+        try:
+            previous_policy = self.policies[-2]
+        except IndexError:
+            previous_policy = {}
+        new_policy = self.policies[-1]
+        newly_added_permissions = find_newly_added_permissions(
+            previous_policy.get("Policy", {}), new_policy.get("Policy", {})
+        )
+        new_no_repo_permissions = {}
+        current_time = int(time.time())
+        for permission, expiration in self.no_repo_permissions.items():
+            if current_time > expiration:
+                self.no_repo_permissions.pop(permission)
+
+        expire_time = current_time + self._no_repo_secs
+        for permission in newly_added_permissions:
+            new_no_repo_permissions[permission] = expire_time
+
+    def _update_opt_out(self) -> None:
+        if self.opt_out and int(self.opt_out["expire"]) < int(time.time()):
+            self.opt_out = {}
+
+    def _update_refreshed(self) -> None:
+        self.refreshed = datetime.datetime.utcnow().isoformat()
 
     def update(self, values: Dict[str, Any], store: bool = True) -> None:
         self._dirty = True
@@ -142,15 +239,20 @@ class Role(BaseModel):
             self.update(role_data, store=False)
             self._updated_fields - set(role_data.keys())
 
+    def mark_inactive(self, store: bool = True) -> None:
+        self.active = False
+        self.store(fields=["Active"])
+
     def store(self, fields: Optional[List[str]] = None) -> None:
         create = False
         try:
-            remote_role_data = get_role_by_id(self.role_id, fields=["LastUpdated"])
-            remote_dt = remote_role_data.get("LastUpdated", "")
-            remote_last_updated = datetime.datetime.strptime(
-                remote_dt, "%Y-%m-%d %H:%M"
-            )
-            if remote_last_updated > self.last_updated:
+            remote_role_data = Role(role_id=self.role_id)
+            remote_role_data.fetch(fields=["LastUpdated"])
+            if (
+                remote_role_data.last_updated
+                and self.last_updated
+                and remote_role_data.last_updated > self.last_updated
+            ):
                 raise IntegrityError("stored role has been updated since last fetch")
         except RoleNotFoundError:
             create = True
@@ -183,6 +285,58 @@ class Role(BaseModel):
                 )
                 self._updated_fields = set()
         self._dirty = False
+
+    def update_role_data(
+        self,
+        current_policy: Dict[str, Any],
+        hooks: RepokidHooks,
+        config: Optional[RepokidConfig] = None,
+        source: str = "Scan",
+        add_no_repo: bool = True,
+        store: bool = True,
+    ) -> None:
+        config = config or CONFIG
+        self.fetch_aa_data(config=config)
+        self.add_policy_version(current_policy, source=source, store=False)
+        if add_no_repo:
+            self._update_no_repo_permissions()
+        self._update_opt_out()
+        self._update_refreshed()
+        minimum_age = config["filter_config"]["AgeFilter"]["minimum_age"]
+        self._calculate_repo_scores(minimum_age, hooks)
+        self.update_stats(source=source, store=False)
+        if store:
+            self.store()
+
+    def update_stats(self, source: str = "Scan", store: bool = True) -> None:
+        new_stats = {
+            "Date": datetime.datetime.utcnow().isoformat(),
+            "DisqualifiedBy": self.disqualified_by,
+            "PermissionsCount": self.total_permissions,
+            "RepoablePermissionsCount": self.repoable_permissions,
+            "Source": source,
+        }
+        try:
+            cur_stats = self.stats[-1]
+        except IndexError:
+            cur_stats = {
+                "DisqualifiedBy": [],
+                "PermissionsCount": 0,
+                "RepoablePermissionsCount": 0,
+            }
+
+        check_fields = [
+            "DisqualifiedBy",
+            "PermissionsCount",
+            "RepoablePermissionsCount",
+        ]
+        changed = any(
+            [new_stats.get(item) != cur_stats.get(item) for item in check_fields]
+        )
+        if changed:
+            self.stats.append(new_stats)
+        if store:
+            self.store(fields=["stats"])
 
 
 class RoleList(object):
@@ -231,6 +385,18 @@ class RoleList(object):
         else:
             raise StopIteration
 
+    @classmethod
+    def from_ids(
+        cls,
+        id_list: Iterable[str],
+        fetch: bool = True,
+        fields: Optional[List[str]] = None,
+    ) -> RoleList:
+        role_list = cls([Role(role_id=role_id) for role_id in id_list])
+        if fetch:
+            map(lambda r: r.fetch(fields=fields), role_list)
+        return role_list
+
     def append(self, role: Role) -> None:
         if not isinstance(role, Role):
             raise AttributeError("cannot add non-Role to RoleList")
@@ -263,3 +429,11 @@ class RoleList(object):
         for arg, value in kwargs.items():
             roles = [role for role in roles if getattr(role, arg, None) == value]
         return RoleList(roles)
+
+    def store(self, fields: Optional[List[str]] = None) -> None:
+        for role in self.roles:
+            role.store(fields=fields)
+
+    def update_stats(self, source: str = "Scan", store: bool = True) -> None:
+        for role in self.roles:
+            role.update_stats(source=source, store=store)
