+++ repokid/utils/iam.py
@@ -27,14 +27,12 @@ from cloudaux.aws.sts import boto3_cached_conn
 from mypy_boto3_dynamodb.service_resource import Table
 from mypy_boto3_iam.client import IAMClient
 
+import repokid.utils.permissions
 from repokid.exceptions import IAMError
 from repokid.role import Role
 from repokid.types import RepokidConfig
 from repokid.types import RepokidHooks
-from repokid.utils import roledata as roledata
-from repokid.utils.dynamo import set_role_data
 from repokid.utils.logging import log_deleted_and_repoed_policies
-from repokid.utils.roledata import partial_update_role_data
 
 LOGGER = logging.getLogger("repokid")
 MAX_AWS_POLICY_SIZE = 10240
@@ -172,7 +170,10 @@ def remove_permissions_from_role(
     Returns:
         None
     """
-    repoed_policies, deleted_policy_names = roledata._get_repoed_policy(
+    (
+        repoed_policies,
+        deleted_policy_names,
+    ) = repokid.utils.permissions._get_repoed_policy(
         role.policies[-1]["Policy"], set(permissions)
     )
 
@@ -205,21 +206,12 @@ def remove_permissions_from_role(
             LOGGER.error(e)
 
     current_policies = get_role_inline_policies(role.dict(), **conn) or {}
-    roledata.add_new_policy_version(dynamo_table, role, current_policies, "Repo")
+    role.add_policy_version(current_policies, "Repo")
 
-    set_role_data(
-        dynamo_table, role.role_id, {"Repoed": datetime.datetime.utcnow().isoformat()}
-    )
+    role.repoed = datetime.datetime.utcnow().isoformat()
     update_repoed_description(role.role_name, conn)
-    partial_update_role_data(
-        role,
-        dynamo_table,
-        account_number,
-        config,
-        conn,
-        hooks,
-        source="ManualPermissionRepo",
-        add_no_repo=False,
+    role.update_role_data(
+        current_policies, hooks, source="ManualPermissionRepo", add_no_repo=False
     )
     LOGGER.info(
         "Successfully removed {permissions} from role: {role} in account {account_number}".format(
